+++ gulpfile.js
@@ -8,13 +8,8 @@ var realFavicon = require ('gulp-real-favicon');
 var fs = require('fs');
 var FAVICON_DATA_FILE = 'faviconData.json';
 var babel = require('gulp-babel');
-
-gulp.task('babel', function () {
-  return gulp.src("./js/script.js")
-    .pipe(babel())
-    .pipe(concat("main.js"))
-    .pipe(gulp.dest("./js"));
-});
+var browserify = require('browserify');
+var source = require('vinyl-source-stream');
 
 gulp.task('less', function(){
 	return gulp.src(['./style/less/*.less'])
@@ -32,6 +27,24 @@ gulp.task('css', function () {
          .pipe(browserSync.stream());
 });
 
+
+gulp.task('browserify', function(){
+  return browserify('./js/script.js')
+    .bundle()
+    .pipe(source('main.js'))
+    .pipe(gulp.dest('./js'))
+    .pipe(browserSync.stream());
+});
+
+gulp.task('babel', function () {
+  return gulp.src("./js/main.js")
+    .pipe(babel())
+    .pipe(concat("main.js"))
+    .pipe(gulp.dest("./js"))
+    .pipe(browserSync.stream());
+});
+
+
 gulp.task('default', gulp.parallel('less', 'css', function(){
        
        browserSync.init({
@@ -40,10 +53,9 @@ gulp.task('default', gulp.parallel('less', 'css', function(){
 
 	  gulp.watch('./index.html').on('change', browserSync.reload);
     gulp.watch('./style/main.css').on('change', browserSync.reload);
-    gulp.watch('./js/*.js').on('change', browserSync.reload);
+    gulp.watch('./js/main.js').on('change', browserSync.reload);
     gulp.watch('./style/less/*.less', gulp.series('less'));
     gulp.watch('./style/dest/*.css', gulp.series('css'));
-
 }));
