+++ index.html
@@ -68,11 +68,11 @@
             <div class="review-user">
                 <span class="review-user__headline">Представьтесь:</span>
                 <label class="review-user__label" for="name" id="name-span">Имя*:</label>
-                <input type="text" class="review-user__input" name="name" id="name-input" required>
+                <input type="text" class="review-user__input" name="name" id="name" required>
                 <label class="review-user__label" for="surname">Фамилия*:</label>
-                <input type="text" class="review-user__input" name="name" id="surname" required>
+                <input type="text" class="review-user__input" name="surname" id="surname" required>
                 <label class="review-user__label" for="patronymic">Отчество:</label>
-                <input type="text" class="review-user__input" name="name" id="patronymic">
+                <input type="text" class="review-user__input" name="patronymic" id="patronymic">
             </div>
         <div class="container">
             <span class="container__headline">Контактная информация:</span>
@@ -82,12 +82,12 @@
                 <span class="wrong-number">Ошибка, перепроверьте номер</span>
                 <label class="users__label" id="label-tel" for="tel">Телефон*:</label>
                 <label class="users__label label-tablet" for="tel">Контактный Телефон*:</label>
-                <input type="tel" class="user-contacts__input" name="tel" pattern="+[0-9]{3}-[0-9]{3}" id="tel" placeholder="Введите телефон" required>
+                <input type="tel" class="user-contacts__input" name="tel" pattern="+[0-9]{3}-[0-9]{3}" id="tel" placeholder="Введите телефон" name ="telephone" required>
             </div>
             <div class="user-contacts__email">
                 <span class="wrong-email">Ошибка, перепроверьте e-mail</span>
                 <label class="users__label label-email" for="email">Электронная почта*:</label>
-                <input type="email" id="email" pattern=".+@globex.com" required class="user-contacts__input" placeholder="Введите e-mail" required>
+                <input type="email" id="email" class="user-contacts__input" placeholder="Введите e-mail" name="email" required>
             </div>
         </div>
         <div class="container">
@@ -107,7 +107,7 @@
             <span class="container__headline">Опишите свои эмоции:</span> </div>
         <textarea class="textarea" placeholder="Опишите подробно все свои восторги" rows="6"></textarea>
         <div class="submit">
-            <button class="push">Отправить отзыв</button>
+            <button type="submit" class="push">Отправить отзыв</button>
             <span class="submit__text">* — обязательные поля</span>
         </div>
     </form>
@@ -127,7 +127,7 @@
             <span class="producer__text">Разработано</span>
         </div>
     </footer>
-    <script data-main="js/main.js" src="js/require.js"></script>
+    <script src="js/main.js"></script>
 </body>
 
 </html>
