+++ js/script.js
@@ -5,9 +5,7 @@ window.onload = function(){
   let closeMobMenu = document.querySelector('.close-mobile-nav');
   let navigation = document.querySelector('.header__nav');
 
-  let mainForm = document.querySelector('.main-form');
-  let emailUser = document.querySelector('.email');
-  let telUser = document.querySelector('.tel');
+  var validate = require("validate.js");
 
   mobileMenuIcon.onclick = () => {
     mobileMenu.style.display = "block";
@@ -23,4 +21,19 @@ window.onload = function(){
     navigation.style.display = "block";
   };
 
+  let form = document.querySelector(".main-form");
+  form.addEventListener('submit', event =>{
+    event.preventDefault();
+
+     const formData = {
+      name: form.name.value,
+      surname: form.surname.value,
+      email: form.email.value,
+      tel: form.tel.value
+     }
+     validate.collectFormValues(formData);
+
+     console.log(formData, validate.collectFormValues(form));
+  });
+ 
 }
