+++ package.json
@@ -20,12 +20,14 @@
     "@babel/core": "^7.4.3",
     "babel-upgrade": "^1.0.1",
     "browser-sync": "^2.26.3",
+    "browserify": "^16.2.3",
     "gulp": "^4.0.0",
     "gulp-less": "^4.0.1",
     "gulp-order": "^1.2.0",
     "normalize.css": "^8.0.1",
     "npm": "^6.9.0",
-    "validate.js": "^0.12.0"
+    "validate.js": "^0.12.0",
+    "vinyl-source-stream": "^2.0.0"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.4.3",
