+++ style/dest/main.css
@@ -463,16 +463,16 @@ body {
     order: 2;
   }
 }
-.review-user #name-input {
+.review-user #name {
   width: calc(100% - 36%);
 }
 @media (max-width: 768px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 41%);
   }
 }
 @media (max-width: 639px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 36%);
     min-height: 56px;
   }
@@ -826,6 +826,9 @@ body {
     max-height: 33vh;
     flex-wrap: wrap;
   }
+}
+.has-error {
+  display: block;
 }
 .user-contacts__number {
   width: 42%;
@@ -1621,16 +1624,16 @@ body {
     order: 2;
   }
 }
-.review-user #name-input {
+.review-user #name {
   width: calc(100% - 36%);
 }
 @media (max-width: 768px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 41%);
   }
 }
 @media (max-width: 639px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 36%);
     min-height: 56px;
   }
@@ -1984,6 +1987,9 @@ body {
     max-height: 33vh;
     flex-wrap: wrap;
   }
+}
+.has-error {
+  display: block;
 }
 .user-contacts__number {
   width: 42%;
@@ -3175,6 +3181,9 @@ body {
     flex-wrap: wrap;
   }
 }
+.has-error {
+  display: block;
+}
 .user-contacts__number {
   width: 42%;
   position: relative;
@@ -3967,16 +3976,16 @@ body {
     order: 2;
   }
 }
-.review-user #name-input {
+.review-user #name {
   width: calc(100% - 36%);
 }
 @media (max-width: 768px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 41%);
   }
 }
 @media (max-width: 639px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 36%);
     min-height: 56px;
   }
@@ -4543,16 +4552,16 @@ body {
     order: 2;
   }
 }
-.review-user #name-input {
+.review-user #name {
   width: calc(100% - 36%);
 }
 @media (max-width: 768px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 41%);
   }
 }
 @media (max-width: 639px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 36%);
     min-height: 56px;
   }
@@ -4906,6 +4915,9 @@ body {
     max-height: 33vh;
     flex-wrap: wrap;
   }
+}
+.has-error {
+  display: block;
 }
 .user-contacts__number {
   width: 42%;
@@ -5903,16 +5915,16 @@ body {
     order: 2;
   }
 }
-.review-user #name-input {
+.review-user #name {
   width: calc(100% - 36%);
 }
 @media (max-width: 768px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 41%);
   }
 }
 @media (max-width: 639px) {
-  .review-user #name-input {
+  .review-user #name {
     width: calc(100% - 36%);
     min-height: 56px;
   }
@@ -6402,6 +6414,9 @@ body {
     max-height: 33vh;
     flex-wrap: wrap;
   }
+}
+.has-error {
+  display: block;
 }
 .user-contacts__number {
   width: 42%;
