+++ style/less/main-first-block.less
@@ -236,7 +236,7 @@
         order: 2;
     });
 
-    #name-input {
+    #name {
         width: calc(100% - 36%);
 
         .tablet-new( {
