+++ style/less/main-second-block.less
@@ -154,6 +154,10 @@
     });
 }
 
+.has-error{
+    display: block;
+}
+
 .user-contacts__number {
     width: 42%;
     .wrong-number {
